Refactor tipizzazione - fix Thunk usage
